requirements.txt
Django==4.2
djangorestframework==3.14.0
django-environ==0.4.5
psycopg2-binary>=2.9
Pillow==10.0.0
django-model-utils==4.3.1
